+++ projects/jsf/src/lib/jsf.module.ts
@@ -26,12 +26,10 @@ import { TableModalService } from './form-content/form-controls/table-modal/tabl
 import { TemplateComponent } from './form-content/form-controls/template/template.component';
 import { TextAreaComponent } from './form-content/form-controls/text-area/text-area.component';
 import { TextComponent } from './form-content/form-controls/text/text.component';
-import { OneOfDropdownComponent } from './form-content/one-of/one-of-dropdown/one-of-dropdown.component';
 import { OneOfComponent } from './form-content/one-of/one-of.component';
 import { SectionComponent } from './form-content/section/section.component';
 import { TabComponent } from './form-content/tabs/tab/tab.component';
 import { TabsComponent } from './form-content/tabs/tabs.component';
-import { TabstripComponent } from './form-content/tabs/tabstrip/tabstrip.component';
 import { FormDataItemService } from './form-data-item.service';
 import { FormService } from './form.service';
 import { JSFComponent } from './jsf.component';
@@ -53,7 +51,6 @@ const tableComponents = [
 const components = [
   TabsComponent,
   TabComponent,
-  TabstripComponent,
   CheckboxComponent,
   FormControlComponent,
   CheckboxWithChildrenComponent,
@@ -64,7 +61,6 @@ const components = [
   FormContentComponent,
   SectionComponent,
   OneOfComponent,
-  OneOfDropdownComponent,
   RadioButtonComponent,
   TextComponent,
   TemplateComponent,